requirements.txt
asgiref==3.5.0
Django==4.0.1
django-admin-interface==0.18.5
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-ordered-model==3.5
django-solo==2.0.0
Pillow==9.0.0
psycopg2==2.9.3
python-dotenv==0.19.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5

sitemaps~=0.1.0